Hitbox adjusted accordingly.

+++ src/utils/HelpMethods.java
@@ -4,7 +4,7 @@ import main.Game;
 
 public class HelpMethods {
 
-    public static boolean CanMoveHere(float x, float y, int width, int height, int[][] levelData) {
+    public static boolean CanMoveHere(float x, float y, float width, float height, int[][] levelData) {
         if (!IsSolid(x, y, levelData))
             if (!IsSolid(x + width, y + height, levelData))
                 if (!IsSolid(x + width, y, levelData))
